Add files via upload
The script reads csv file exported from the Beagle sniffer and looks for Get PMBus Readings command. When it's completion code equals 00 then the temperature reading is collected from the response as data sheet for Infineon PXE-1410 documents. It also shows the temperature value when the exponent part of the reading is not taken into account due to suspicion it might lead to wrong values reported by BMC. The scripts reports the number of non zero competition code per each cc. The temperature histogram is presented too.